feat: 팔로잉 피드 API 구현
Former-commit-id: 79dfaa3739a2850f8849e24b13dec38d6be12246

+++ back/src/main/java/com/ssafy/recode/domain/feed/dto/response/ProblemDto.java
@@ -8,11 +8,11 @@ import lombok.Getter;
 @Getter
 @Builder
 public class ProblemDto {
-    private Long problemId;
+    private Integer problemId;
     private String problemName;
     private Integer problemTier;
     @Builder
-    public ProblemDto(Long problemId, String problemName, Integer problemTier) {
+    public ProblemDto(Integer problemId, String problemName, Integer problemTier) {
         this.problemId = problemId;
         this.problemName = problemName;
         this.problemTier = problemTier;
